Create hidescreens.rpy
In order not to ruin all the addressing and other details that are unimportant to us, in Renpy you cannot make a call from the screen (go to the label with the subsequent return). You can go to the label in a new context, but then Renpy will hide all the screens. This is an example of how you can organize a call by a button and display hidden screens, as if they were not hidden. But the button on which we jumped to the mark is better to hide. Or replace the back button.