Create visualeditor.rpy
If you are tired of writing code to build a level for a mini-game, then you will want to create a visual environment for yourself, drag and drop objects on the screen and save the results to a file. Here you will see an example and on its basis you can create your own level editors, characters, anything. In order not to bother with formats for saving data, I created a module that writes any classes, their lists, dictionaries, or simply variables in the form of python code to a file with the .pry extension. It is necessary to specify the file name (without extension), the name of the function that will be responsible for the “download”, its “level” (the execution order of the init block) and the names of all variables, lists, etc. that need to be saved, separated by a comma,