Create sayboxcolor.rpy
Just changing the background of the seibox during the game is not enough. Because the next time you start the game and load the save, the background will be reset to the one that was by default. The easiest way out is a picture that depends on the variable. And the variables are saved and loaded along with the rest of the data.